move query/common to query_common crate
