+++ query/common/Cargo.toml
@@ -1,0 +1,81 @@
+[package]
+name = "query-common"
+version = "0.1.0"
+description = "A real-time Cloud Distributed Query Engine"
+authors = ["Databend Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[lib]
+doctest = false
+test = false
+
+[features]
+default = ["simd"]
+simd = ["common-arrow/simd"]
+tokio-console = ["common-tracing/console", "common-base/tracing"]
+
+[dependencies]
+# Workspace dependencies
+common-arrow = { path = "../../common/arrow" }
+common-base = { path = "../../common/base" }
+common-clickhouse-srv = { path = "../../common/clickhouse-srv" }
+common-datavalues = { path = "../../common/datavalues" }
+common-exception = { path = "../../common/exception" }
+common-grpc = { path = "../../common/grpc" }
+common-infallible = { path = "../../common/infallible" }
+common-io = { path = "../../common/io" }
+common-macros = { path = "../../common/macros" }
+common-mem-allocator = { path = "../../common/mem-allocator" }
+common-meta-api = { path = "../../common/meta/api" }
+common-meta-embedded = { path = "../../common/meta/embedded" }
+common-meta-grpc = { path = "../../common/meta/grpc" }
+common-meta-raft-store = { path = "../../common/meta/raft-store" }
+common-meta-sled-store = { path = "../../common/meta/sled-store" }
+common-meta-types = { path = "../../common/meta/types" }
+common-tracing = { path = "../../common/tracing" }
+
+# Github dependencies
+bincode = { git = "https://github.com/datafuse-extras/bincode", rev = "fd3f9ff" }
+sqlparser = { git = "https://github.com/datafuse-extras/sqlparser-rs", rev = "472f5b6" }
+
+# Crates.io dependencies
+ahash = "0.7.6"
+async-stream = "0.3.2"
+async-trait = "0.1.52"
+base64 = "0.13.0"
+byteorder = "1.4.3"
+bytes = "1.1.0"
+chrono = "0.4.19"
+chrono-tz = "0.6.1"
+clap = { version = "3.1.3", features = ["derive", "env"] }
+dyn-clone = "1.0.4"
+futures = "0.3.21"
+headers = "0.3.7"
+http = "0.2.6"
+hyper = "0.14.17"
+jwt-simple = "0.10.8"
+num = "0.4.0"
+num_cpus = "1.13.1"
+octocrab = "0.15.4"
+once_cell = "1.9.0"
+opendal = "0.2.0"
+parquet-format-async-temp = "0.2.0"
+paste = "1.0.6"
+poem = { version = "1.3.6", features = ["rustls", "multipart"] }
+prost = "0.9.0"
+rand = "0.8.5"
+regex = "1.5.5"
+reqwest = "0.11.9"
+rsa = "0.5.0"
+serde = { version = "1.0.136", features = ["derive"] }
+serde_json = "1.0.79"
+sha1 = "0.10.1"
+sha2 = "0.10.2"
+threadpool = "1.8.1"
+tokio-rustls = "0.23.2"
+tokio-stream = { version = "0.1.8", features = ["net"] }
+tonic = "0.6.2"
+typetag = "0.1.5"
+uuid = { version = "0.8.2", features = ["serde", "v4"] }
